// web/src/lib/djen-types.gen.ts
export interface paths {
    "/api/v1/login": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** @description Método de autenticação, para operações de inclusão e remoção de comunicações processuais. Método a ser utilizado pelos Tribunais. */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            /** @description Login do usuário cadastrado no sistema de controle de acesso do CNJ (SCA) */
            requestBody?: {
                content: {
                    "application/json": {
                        login: string;
                        senha: string;
                    };
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            user?: {
                                /** @description Id do usuário */
                                id?: number;
                                nome?: string;
                                email?: string;
                                cpf?: string;
                            };
                            /** @description Token de autenticação a ser utilizado nas próximas requisições */
                            access_token?: string;
                        };
                    };
                };
                /** @description Credenciais inválidas */
                403: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description Erro negocial */
                422: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/api/v1/comunicacao/{hash}/certidao": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description GET /api/v1/comunicacao/{hash}/certidao */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    hash: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/comunicacao/tribunal": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Este endpoint retorna lista de tribunais por UF de atuação com as datas de último envio disponibilizado pelo tribunal. */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Sucesso ao retornar as informações dos Tribunais */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @description Identificador único do Tribunal */
                            id?: number;
                            /** @description Nome completo do Tribunal */
                            nome?: string;
                            /** @description Sigla do Tribunal */
                            sigla?: string;
                            /** @description Descrição da jurisdição do Tribunal */
                            jurisdicao?: string;
                            /** @description Endereço completo do Tribunal */
                            endereco?: string;
                            /** @description Telefone de contato do Tribunal */
                            telefone?: string;
                        }[];
                    };
                };
                /** @description Nenhum Tribunal encontrado */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example error */
                            status?: string;
                            /** @example Nenhum Tribunal encontrado. */
                            message?: string;
                        };
                    };
                };
                /** @description Erro interno no servidor */
                500: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example error */
                            status?: string;
                            /** @example Erro interno ao buscar as informações dos Tribunais. */
                            message?: string;
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/comunicacao/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post?: never;
        /** @description Cancela comunicações. Caso a comunicação não tenha sido disponibilizada, não aparecerá em qualquer forma nas consultas. Caso a comunicação já tenha sido disponibilizada, o conteúdo será substituído pelo motivo de cancelamento. Pode levar algumas horas para refletir em todas as pesquisas do sistema. */
        delete: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    id: string;
                };
                cookie?: never;
            };
            /** @description Request body opcional caso a comunicação seja cancelada antes da disponibilização e obrigatória após a disponibilização. */
            requestBody?: {
                content: {
                    "application/json": {
                        motivo_cancelamento: string;
                    };
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @description Indica status da requisição. Ex. success */
                            status?: string;
                            /** @description Mensagem de confirmação. */
                            message?: string;
                        };
                    };
                };
                /** @description Credenciais inválidas */
                403: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description Erro negocial */
                422: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/caderno/{sigla_tribunal}/{data}/{meio}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Método para download dos cadernos compactados de comunicações de cada tribunal.<br>O endpoint retorna metadados sobre o caderno e URL temporária (5 minutos) para download.<br>Os cadernos do dia atual são disponibilizados a partir das 02:00. */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    sigla_tribunal: string;
                    data: string;
                    /** @description "E" para Edital e "D" para Diário Eletrônico */
                    meio: "E" | "D";
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @description Nome completo do tribunal */
                            tribunal?: string;
                            /** @description Sigla do tribunal */
                            sigla_tribunal?: string;
                            /** @description D ou E */
                            meio?: string;
                            /** @description Status do processamento. Não Processado, Em processamento, Processado, Sem comunicações ou Cancelado. */
                            status?: string;
                            /** @description Número da versão do diário. A versão é incrementada quando o diário precisa ser reprocessado, como no caso de cancelamento de comunicações. */
                            versao?: string;
                            /** @description Data referência no formado YYYY-MM-DD */
                            data?: string;
                            /** @description Número de comunicações */
                            total_comunicacoes?: number;
                            /** @description Número de arquivos dentro do pacote compactado */
                            numero_paginas?: number;
                            /** @description Hash SHA256 do arquivo */
                            hash?: string;
                            /** @description URL temporária de download */
                            url?: string;
                        };
                    };
                };
                /** @description Erro negocial */
                422: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}