// web/src/lib/djen-zod.gen.ts
export const getApiV1ComunicacaoResponse = zod.object({
  "status": zod.string().optional().describe('Status da operação'),
  "message": zod.string().optional(),
  "count": zod.number().optional().describe('Quantidade de comunicações retornadas'),
  "items": zod.array(zod.object({
  "id": zod.number().optional().describe('Id da comunicação'),
  "data_disponibilizacao": zod.string().optional().describe('Data em que a comunicação foi disponibilizada'),
  "siglaTribunal": zod.string().optional(),
  "tipoComunicacao": zod.string().optional().describe('Citação, intimação, edital, pauta de julgamento, lista de distribuição ou ata de sessão'),
  "nomeOrgao": zod.string().optional(),
  "texto": zod.string().optional().describe('Teor da Comunicação'),
  "numero_processo": zod.string().optional(),
  "meio": zod.string().optional().describe('Plataforma de Editais ou Diário de Justiça Eletrônico'),
  "link": zod.string().optional().describe('Link para o inteiro teor da comunicação'),
  "tipoDocumento": zod.string().optional(),
  "nomeClasse": zod.string().optional().describe('Classe do processo'),
  "codigoClasse": zod.string().optional(),
  "numeroComunicacao": zod.number().optional().describe('Número de identificação da comunicação'),
  "ativo": zod.boolean().optional(),
  "hash": zod.string().optional(),
  "datadisponibilizacao": zod.string().optional(),
  "meiocompleto": zod.string().optional(),
  "numeroprocessocommascara": zod.string().optional(),
  "destinatarios": zod.array(zod.object({
  "nome": zod.string().optional(),
  "polo": zod.string().optional().describe('A para Ativo, P para Passivo, T para Terceiro interessado e D para Outros Destinatários'),
  "comunicacao_id": zod.number().optional()
})).optional(),
  "destinatarioadvogados": zod.array(zod.object({
  "id": zod.number().optional(),
  "comunicacao_id": zod.number().optional(),
  "advogado_id": zod.number().optional(),
  "created_at": zod.string().optional(),
  "updated_at": zod.string().optional(),
  "advogado": zod.object({
  "id": zod.number(),
  "nome": zod.string(),
  "numero_oab": zod.string(),
  "uf_oab": zod.string()
}).optional()
})).optional()
})).optional()
})